requirements.txt
langchain
langchain-openai
langchain-community
chromadb
tavily-python
requests
python-dotenv
streamlit
